refactor(backend): remove unuse package and format upload router
