+++ backend/src/controllers/orderController.ts
@@ -15,12 +15,6 @@ class OrderController {
       res.status(400);
       throw new Error("No order items");
     } else {
-      // NOTE: here we must assume that the prices from our client are incorrect.
-      // We must only trust the price of the item as it exists in
-      // our DB. This prevents a user paying whatever they want by hacking our client
-      // side code - https://gist.github.com/bushblade/725780e6043eaf59415fbaf6ca7376ff
-
-      // get the ordered items from our database
       const itemsFromDB: IOrderItem[] = await Product.find({
         _id: { $in: orderItems.map((x: any) => x._id) },
       });
