+++ backend/src/middleware/checkObjectId.ts
@@ -1,14 +1,6 @@
 import { isValidObjectId } from 'mongoose';
 import { Request, Response, NextFunction } from "express";
 
-/**
- * Checks if the req.params.id is a valid Mongoose ObjectId.
- *
- * @param {import('express').Request} req - The Express request object.
- * @param {import('express').Response} res - The Express response object.
- * @param {import('express').NextFunction} next - The Express next middleware function.
- * @throws {Error} Throws an error if the ObjectId is invalid.
- */
 
 function checkObjectId(req: Request, res: Response, next: NextFunction) {
   if (!isValidObjectId(req.params.id)) {
