+++ package.json
@@ -18,8 +18,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@tensorflow/tfjs": "^4.14.0",
-    "@tensorflow/tfjs-node": "^4.14.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.20.2",
     "colors": "^1.4.0",
@@ -29,8 +27,7 @@
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.6.3",
     "multer": "^1.4.5-lts.1",
-    "natural": "^6.10.4",
-    "node-gyp": "^10.0.1"
+    "natural": "^6.10.4"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
